+++ packages/genesys/packages/react-native-components/src/input/Button/hooks/useButtonStyles.ts
@@ -1,47 +1,108 @@
-import { TextStyle, ViewStyle } from "react-native";
+import useDefaultStyles from "./useButtonDefaultStyle.ts";
 import { extractTextStyles } from "@peersyst/react-native-utils";
 import { ButtonRootStyle, ButtonStyle, ButtonTextStyle } from "../Button.types";
 import { useMemo } from "react";
-import useButtonDefaultStyle from "./useButtonDefaultStyle.ts";
-import { useGlobalStyles } from "../../../config";
-import { StyleFlags } from "../../../utils";
-import { useMergeStyles } from "../../../hooks";
+import { ButtonSize, ButtonVariant } from "@peersyst/react-components-core";
+import { useGlobalStyles } from "src/config";
 
 export interface UseButtonStylesResult {
-    textStyle: TextStyle;
-    rootStyle: Omit<ButtonStyle, keyof TextStyle> & ViewStyle;
+    textStyle: ButtonTextStyle;
+    rootStyle: ButtonRootStyle;
 }
 
+// TODO: Refactor or add sizes merge to stateful styles
 export default function useButtonStyles(
     style: ButtonStyle,
+    variant: ButtonVariant,
+    size: ButtonSize,
+    disabled: boolean,
+    pressed: boolean,
     color: string | undefined,
-    {
-        filled,
-        outlined,
-        text,
-        sm,
-        md,
-        lg,
-        pressed,
-        disabled,
-    }: StyleFlags<
-        ButtonStyle,
-        "filled" | "outlined" | "text" | "sm" | "md" | "lg" | "pressed" | "disabled"
-    >,
-): [ButtonRootStyle, ButtonTextStyle] {
-    const defaultStyle = useButtonDefaultStyle(color);
-    const globalStyle = useGlobalStyles("Button");
+): UseButtonStylesResult {
+    const { defaultStyles, defaultDisabledStyles, defaultPressedStyles, defaultSizeStyles } =
+        useDefaultStyles(color);
+    const {
+        disabled: globalDisabledStyles = {},
+        pressed: globalPressedStyles = {},
+        ...globalStyles
+    } = useGlobalStyles("Button");
+    const { disabled: disabledStyles = {}, pressed: pressedStyles = {}, ...styles } = style;
 
-    const mergedStyles = useMergeStyles([defaultStyle, globalStyle, style], {
-        filled,
-        outlined,
-        text,
-        sm,
-        md,
-        lg,
-        pressed,
-        disabled,
-    });
+    const [textStyles, rootStyles] = useMemo(
+        () => extractTextStyles({ ...defaultStyles, ...globalStyles, ...styles }),
+        [defaultStyles, globalStyles, styles],
+    );
 
-    return useMemo(() => extractTextStyles(mergedStyles), [mergedStyles]);
+    const [variantTextStyles, variantRootStyles] = useMemo(
+        () =>
+            extractTextStyles({
+                ...defaultStyles[variant],
+                ...globalStyles[variant],
+                ...styles[variant],
+            }),
+        [defaultStyles, globalStyles, styles, variant],
+    );
+
+    const [sizeTextStyles, sizeRootStyles] = useMemo(
+        () =>
+            extractTextStyles({
+                ...defaultSizeStyles[size],
+                ...globalStyles[size],
+                ...styles[size],
+            }),
+        [defaultSizeStyles, globalStyles, styles, size],
+    );
+
+    const [pressedTextStyles, pressedRootStyles] = useMemo(
+        () =>
+            extractTextStyles({
+                ...defaultPressedStyles,
+                ...globalPressedStyles,
+                ...pressedStyles,
+            }),
+        [defaultPressedStyles, globalPressedStyles, pressedStyles],
+    );
+
+    const [variantPressedTextStyles, variantPressedRootStyles] = useMemo(() => {
+        return extractTextStyles({
+            ...defaultPressedStyles[variant],
+            ...globalPressedStyles[variant],
+            ...pressedStyles[variant],
+        });
+    }, [defaultPressedStyles, globalPressedStyles, pressedStyles, variant]);
+
+    const [disabledTextStyles, disabledRootStyles] = useMemo(
+        () =>
+            extractTextStyles({
+                ...defaultDisabledStyles,
+                ...globalDisabledStyles,
+                ...disabledStyles,
+            }),
+        [defaultDisabledStyles, globalDisabledStyles, disabledStyles],
+    );
+
+    const [variantDisabledTextStyles, variantDisabledRootStyles] = useMemo(() => {
+        return extractTextStyles({
+            ...defaultDisabledStyles[variant],
+            ...globalDisabledStyles[variant],
+            ...disabledStyles[variant],
+        });
+    }, [defaultDisabledStyles, globalDisabledStyles, disabledStyles, variant]);
+
+    const textStyle = {
+        ...textStyles,
+        ...variantTextStyles,
+        ...sizeTextStyles,
+        ...(pressed && { ...pressedTextStyles, ...variantPressedTextStyles }),
+        ...(disabled && { ...disabledTextStyles, ...variantDisabledTextStyles }),
+    };
+    const rootStyle = {
+        ...rootStyles,
+        ...variantRootStyles,
+        ...sizeRootStyles,
+        ...(disabled && { ...disabledRootStyles, ...variantDisabledRootStyles }),
+        ...(pressed && { ...pressedRootStyles, ...variantPressedRootStyles }),
+    };
+
+    return { textStyle, rootStyle };
 }