+++ packages/react-native-styled/src/styledWithAs.tsx
@@ -41,14 +41,7 @@ export default function styled<
                         ),
                         sxProp?.({ theme, dimensions, safeAreaInsets }),
                     ),
-                [
-                    theme,
-                    dimensions,
-                    safeAreaInsets,
-                    JSON.stringify(styleProp),
-                    JSON.stringify(rest),
-                    sxProp?.toString(),
-                ],
+                [theme, dimensions, safeAreaInsets, styleProp, rest, sxProp?.toString()],
             );
 
             const finalProps = {