+++ packages/genesys/packages/react-native-components/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 2.1.20
+
+[Button] Rollback button styles
+
 ## 2.1.19
 
 [fix] Imports